Split shopping list detail into View and EditMode sub-components
Extracts ShoppingListView and ShoppingListEditMode as focused components
in the same folder. page.tsx becomes a slim orchestrator that manages
isEditMode state and header actions, delegating all rendering to the
appropriate sub-component. React Query deduplicates item fetches via
cache so there is no double-fetching on mode switches.
